Simplify wind direction calculation
The previous implementation attempted to make it trivial for the user by calculating the wind vane resistance and comparing to the values specified by the datasheet, meaning the user doesn't need to punch in a bunch of arbitrary values. However not all boards use a simple voltage divider, and not all ADCs are linear enough, meaning the user would have to punch in arbitrary values anyways.

This new approach removes any conception about the circuit (other than being some kind voltage divider), and simply compares the measured ADC value to expected ADC values for each direction.